backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.1
pgvector==0.3.6
pydantic-settings==2.4.0
httpx==0.27.2
google-generativeai==0.8.3

# Testing dependencies
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.24.0
